feat: add internal cache and logging round tripper
fix: ci build go get

fix: enable verbose build
