+++ .dockerignore
@@ -1,13 +1,7 @@
-draft.toml
-target/classes
-target/generated-sources
-target/generated-test-sources
-target/maven-archiver
-target/maven-status
-target/surefire-reports
-target/test-classes
-target/*.original
+.devcontainer/
+.git/
+.github/
+.gitignore
 charts/
-NOTICE
-LICENSE
-README.md
+deployment/
+draft.toml
