+++ Dockerfile
@@ -4,20 +4,20 @@ LABEL maintainer="Said Sef <saidsef@gmail.com> (saidsef.co.uk/)"
 ENV PORT ${PORT:-8080}
 
 WORKDIR /app
-COPY geocode.go go.mod go.sum /app/
+COPY . /app/
 RUN apk add --no-cache curl git && \
-    go get github.com/prometheus/client_golang/prometheus/promhttp && \
-    go build geocode.go
+    go get -v && \
+    go build -v -o /app/ ./...
 
 FROM docker.io/alpine:3.20
 LABEL maintainer="Said Sef <saidsef@gmail.com> (saidsef.co.uk/)"
 
 USER nobody
 
-COPY --from=builder /app/geocode /usr/bin/
+COPY --from=builder /app/faas-reverse-geocoding /usr/bin/
 
 HEALTHCHECK --interval=60s --timeout=10s CMD curl --fail 'http://localhost:${PORT}/' || exit 1
 
 EXPOSE ${PORT}
 
-CMD ["/usr/bin/geocode"]
+CMD ["/usr/bin/faas-reverse-geocoding"]
